Checkpoint.

+++ src/shaders/deferred_bindings.wgsl
@@ -7,7 +7,7 @@ var normal_texture: texture_2d<f32>;
 @group(1) @binding(2)
 var depth_gradient_texture: texture_2d<f32>;
 @group(1) @binding(3)
-var instance_material_texture: texture_2d<u32>;
+var instance_material_texture: texture_2d<f32>;
 @group(1) @binding(4)
 var velocity_uv_texture: texture_2d<f32>;
 @group(1) @binding(5)
